+++ metricflow/plan_conversion/sql_join_builder.py
@@ -1,6 +1,9 @@
 from dataclasses import dataclass
-from typing import List, Sequence
+from typing import List, Optional, Sequence, TypeVar
 
+from metricflow.dataflow.dataflow_plan import JoinOverTimeRangeNode
+from metricflow.model.objects.metric import CumulativeMetricWindow
+from metricflow.plan_conversion.sql_dataset import SqlDataSet
 from metricflow.sql.sql_plan import SqlExpressionNode, SqlJoinDescription, SqlJoinType, SqlSelectStatementNode
 from metricflow.sql.sql_exprs import (
     SqlColumnReference,
@@ -9,7 +12,129 @@ from metricflow.sql.sql_exprs import (
     SqlComparisonExpression,
     SqlLogicalExpression,
     SqlLogicalOperator,
+    SqlTimeDeltaExpression,
 )
+from metricflow.time.time_granularity import TimeGranularity
+
+SqlDataSetT = TypeVar("SqlDataSetT", bound=SqlDataSet)
+
+
+def _make_cumulative_metric_join_condition(
+    from_data_set_alias: str,
+    from_data_set_metric_time_col: str,
+    right_data_set_alias: str,
+    right_data_set_identifier_col: str,
+    window: Optional[CumulativeMetricWindow],
+) -> SqlExpressionNode:
+    """Creates SqlLogicalExpression representing a cumulative metric self-join condition"""
+    if window is None:  # accumulate over all time
+        return SqlComparisonExpression(
+            left_expr=SqlColumnReferenceExpression(
+                SqlColumnReference(
+                    table_alias=from_data_set_alias,
+                    column_name=from_data_set_metric_time_col,
+                )
+            ),
+            comparison=SqlComparison.LESS_THAN_OR_EQUALS,
+            right_expr=SqlColumnReferenceExpression(
+                SqlColumnReference(
+                    table_alias=right_data_set_alias,
+                    column_name=right_data_set_identifier_col,
+                )
+            ),
+        )
+
+    # ds <= <date> and  ds > <date> - <window>
+    return SqlLogicalExpression(
+        operator=SqlLogicalOperator.AND,
+        args=(
+            SqlComparisonExpression(
+                left_expr=SqlColumnReferenceExpression(
+                    SqlColumnReference(
+                        table_alias=from_data_set_alias,
+                        column_name=from_data_set_metric_time_col,
+                    )
+                ),
+                comparison=SqlComparison.LESS_THAN_OR_EQUALS,
+                right_expr=SqlColumnReferenceExpression(
+                    SqlColumnReference(
+                        table_alias=right_data_set_alias,
+                        column_name=right_data_set_identifier_col,
+                    )
+                ),
+            ),
+            SqlComparisonExpression(
+                left_expr=SqlColumnReferenceExpression(
+                    SqlColumnReference(
+                        table_alias=from_data_set_alias,
+                        column_name=from_data_set_metric_time_col,
+                    )
+                ),
+                comparison=SqlComparison.GREATER_THAN,
+                right_expr=SqlTimeDeltaExpression(
+                    SqlColumnReferenceExpression(
+                        SqlColumnReference(
+                            table_alias=right_data_set_alias,
+                            column_name=right_data_set_identifier_col,
+                        )
+                    ),
+                    count=window.count,
+                    granularity=window.granularity,
+                    grain_to_date=None,
+                ),
+            ),
+        ),
+    )
+
+
+def _make_grain_to_date_cumulative_metric_join_condition(
+    from_data_set_alias: str,
+    from_data_set_metric_time_col: str,
+    right_data_set_alias: str,
+    right_data_set_identifier_col: str,
+    grain_to_date: TimeGranularity,
+) -> SqlLogicalExpression:
+    """Creates SqlLogicalExpression representing a grain_to_date cumulative metric self-join condition"""
+    return SqlLogicalExpression(
+        operator=SqlLogicalOperator.AND,
+        args=(
+            SqlComparisonExpression(
+                left_expr=SqlColumnReferenceExpression(
+                    SqlColumnReference(
+                        table_alias=from_data_set_alias,
+                        column_name=from_data_set_metric_time_col,
+                    )
+                ),
+                comparison=SqlComparison.LESS_THAN_OR_EQUALS,
+                right_expr=SqlColumnReferenceExpression(
+                    SqlColumnReference(
+                        table_alias=right_data_set_alias,
+                        column_name=right_data_set_identifier_col,
+                    )
+                ),
+            ),
+            SqlComparisonExpression(
+                left_expr=SqlColumnReferenceExpression(
+                    SqlColumnReference(
+                        table_alias=from_data_set_alias,
+                        column_name=from_data_set_metric_time_col,
+                    )
+                ),
+                comparison=SqlComparison.GREATER_THAN_OR_EQUALS,
+                right_expr=SqlTimeDeltaExpression(
+                    SqlColumnReferenceExpression(
+                        SqlColumnReference(
+                            table_alias=right_data_set_alias,
+                            column_name=right_data_set_identifier_col,
+                        )
+                    ),
+                    count=0,
+                    granularity=grain_to_date,
+                    grain_to_date=grain_to_date,
+                ),
+            ),
+        ),
+    )
 
 
 @dataclass(frozen=True)
@@ -18,6 +143,23 @@ class ColumnEqualityDescription:
 
     left_column_alias: str
     right_column_alias: str
+
+
+@dataclass(frozen=True)
+class AnnotatedSqlDataSet:
+    """Class to bind a DataSet to transient properties associated with it at a given point in the SqlQueryPlan"""
+
+    data_set: SqlDataSet
+    alias: str
+    _metric_time_column_name: Optional[str]
+
+    @property
+    def metric_time_column_name(self) -> str:
+        """Direct accessor for the optional metric time name, only safe to call when we know that value is set"""
+        assert (
+            self._metric_time_column_name
+        ), "Expected a valid metric time dimension name to be associated with this dataset, but did not get one!"
+        return self._metric_time_column_name
 
 
 class SqlQueryPlanJoinBuilder:
@@ -80,4 +222,57 @@ class SqlQueryPlanJoinBuilder:
             right_source_alias=right_source_alias,
             on_condition=on_condition,
             join_type=join_type,
+        )
+
+    @staticmethod
+    def make_cumulative_metric_time_range_join_description(
+        node: JoinOverTimeRangeNode[SqlDataSetT],
+        metric_data_set: AnnotatedSqlDataSet,
+        time_spine_data_set: AnnotatedSqlDataSet,
+    ) -> SqlJoinDescription:
+        """Make a join description to connect a cumulative metric input to a time spine dataset
+
+        Cumulative metrics must be joined against a time spine in a backward-looking fashion, with
+        a range determined by a time window (delta against metric_time) and optional cumulative grain.
+        """
+        # Temporary local variables to keep join description logic consistent with the original
+        input_data_set = metric_data_set.data_set
+        input_data_set_metric_time_col = metric_data_set.metric_time_column_name
+        input_data_set_alias = metric_data_set.alias
+        time_spine_data_set_alias = time_spine_data_set.alias
+        time_spine_data_set_time_dimension_col = time_spine_data_set.metric_time_column_name
+
+        # Build an expression like "a.ds <= b.ds AND a.ds >= b.ds - <window>
+        # If no window is present we join across all time -> "a.ds <= b.ds"
+        constrain_metric_time_column_condition_both: Optional[SqlExpressionNode] = None
+        if node.window is not None:
+            constrain_metric_time_column_condition_both = _make_cumulative_metric_join_condition(
+                input_data_set_alias,
+                input_data_set_metric_time_col,
+                time_spine_data_set_alias,
+                time_spine_data_set_time_dimension_col,
+                node.window,
+            )
+        elif node.grain_to_date is not None:
+            constrain_metric_time_column_condition_both = _make_grain_to_date_cumulative_metric_join_condition(
+                input_data_set_alias,
+                input_data_set_metric_time_col,
+                time_spine_data_set_alias,
+                time_spine_data_set_time_dimension_col,
+                node.grain_to_date,
+            )
+        elif node.window is None:  # `window is None` for clarity (could be else since we have window and !window)
+            constrain_metric_time_column_condition_both = _make_cumulative_metric_join_condition(
+                input_data_set_alias,
+                input_data_set_metric_time_col,
+                time_spine_data_set_alias,
+                time_spine_data_set_time_dimension_col,
+                None,
+            )
+
+        return SqlJoinDescription(
+            right_source=input_data_set.sql_select_node,
+            right_source_alias=input_data_set_alias,
+            on_condition=constrain_metric_time_column_condition_both,
+            join_type=SqlJoinType.INNER,
         )
